refactor and improve like post

+++ src/main/java/com/t1908e/memeportalapi/repository/PostLikeRepository.java
@@ -6,5 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface PostLikeRepository extends JpaRepository<PostLike, Integer> {
-    Optional<PostLike> findPostLikeByPostIdAndUserId(int postId,int userId);
+    Optional<PostLike> findPostLikeByPostIdAndUserId(int postId,long userId);
+
 }
